refactoring of GitException
