+++ plugins/plugin-bitbucket/codenvy-plugin-bitbucket-ext-bitbucket/src/main/java/org/eclipse/che/ide/ext/bitbucket/server/rest/BitbucketService.java
@@ -17,7 +17,7 @@ package org.eclipse.che.ide.ext.bitbucket.server.rest;
 import org.eclipse.che.api.core.NotFoundException;
 import org.eclipse.che.api.core.ServerException;
 import org.eclipse.che.api.core.UnauthorizedException;
-import org.eclipse.che.api.git.GitException;
+import org.eclipse.che.api.git.exception.GitException;
 import org.eclipse.che.api.ssh.server.SshServiceClient;
 import org.eclipse.che.api.ssh.shared.dto.GenerateSshPairRequest;
 import org.eclipse.che.api.ssh.shared.model.SshPair;
