+++ plugins/plugin-microsoft/codenvy-plugin-microsoft-git-provider/src/main/java/org/eclipse/che/ide/ext/git/server/nativegit/MicrosoftOAuthCredentialsProvider.java
@@ -16,7 +16,7 @@ package org.eclipse.che.ide.ext.git.server.nativegit;
 
 import org.eclipse.che.api.auth.oauth.OAuthTokenProvider;
 import org.eclipse.che.api.auth.shared.dto.OAuthToken;
-import org.eclipse.che.api.git.GitException;
+import org.eclipse.che.api.git.exception.GitException;
 import org.eclipse.che.api.git.shared.ProviderInfo;
 import org.eclipse.che.commons.annotation.Nullable;
 import org.eclipse.che.commons.env.EnvironmentContext;
